+++ trunk/srcpd/src/i2c-dev.h
@@ -22,24 +22,24 @@ typedef I2C_VALUE I2C_DEV_VALUES[MAX_I2C_DEVICES][MAX_MULTIPELEX_BUSSES];
 
 typedef struct _I2CDEV_DATA {
 
-	int number_ga;
-	int multiplex_busses;
-	int ga_hardware_inverters;
-	int ga_reset_devices;    
- 	int ga_min_active_time;
+  int number_ga;
+  int multiplex_busses;
+  int ga_hardware_inverters;
+  int ga_reset_devices;
+  int ga_min_active_time;
 
     I2C_DEV_VALUES i2c_values;
 } I2CDEV_DATA;
 
-int readconfig_I2C_DEV(xmlDocPtr doc, xmlNodePtr node, int busnumber);
-int init_lineI2C_DEV(int );
-int init_bus_I2C_DEV(int );
-int term_bus_I2C_DEV(int );
+int readconfig_I2C_DEV(xmlDocPtr doc, xmlNodePtr node, long int busnumber);
+int init_lineI2C_DEV(long int );
+long int init_bus_I2C_DEV(long int );
+long int term_bus_I2C_DEV(long int );
 int getDescription_I2C_DEV(char *reply);
 void* thr_sendrec_I2C_DEV(void *);
 
 // helper functions
-void reset_ga(int busnumber, int busfd);
-void select_bus(int mult_busnum, int busfd, int busnumber);
+void reset_ga(long int busnumber, int busfd);
+void select_bus(int mult_busnum, int busfd, long int busnumber);
 
 #endif
