+++ trunk/srcpd/src/loopback.h
@@ -9,11 +9,11 @@ typedef struct _LOOPBACK_DATA {
     int number_fb;
 } LOOPBACK_DATA;
 
-int readconfig_LOOPBACK(xmlDocPtr doc, xmlNodePtr node, int busnumber);
-int init_bus_LOOPBACK(int );
-int term_bus_LOOPBACK(int );
-int init_gl_LOOPBACK(struct _GLSTATE *);
-int init_ga_LOOPBACK(struct _GASTATE *);
+int readconfig_LOOPBACK(xmlDocPtr doc, xmlNodePtr node, long int busnumber);
+long int init_bus_LOOPBACK(long );
+long int term_bus_LOOPBACK(long );
+long int init_gl_LOOPBACK(struct _GLSTATE *);
+long int init_ga_LOOPBACK(struct _GASTATE *);
 int getDescription_LOOPBACK(char *reply);
 void* thr_sendrec_LOOPBACK(void *);
